For KDA use an experimental two-way serial protocol.
This will break compile for all files other than kda.inc and it breaks
normal PPM signal functionality.  It also assumes that RCP_IN is
inverted like it is on KDA boards.  This is a dummy serial protocol that
uses two pins (RCP_IN and MOSI) in which the other side can send a
single byte command digitally to set the throttle value and can read
a single byte that contains the number of commutations since the last
time this byte was read.  Seems that currently the ISR takes too long
causing a glitch in the motor's speed by interrupting the main loop for
too long.