Replace serial protocol with a new 14-bit thing.
This breaks anything that uses the old two-pin (PWM in + MOSI) protocol.
The new one is similar, but every write is 14-bit where the low 13 bits
are the desired throttle value (full 13-bit range) and the MSB is used
for reverse mode (1-reverse, 0-normal).
The reads are now 16-bit and the whole commutation counter is under
address 0, while addres 1 is unused (returns 0x1234 so that
communication can be checked).  It is recommended, though that 14-bit
reads are used so that a spurious bit or a missing bit will never cause
an operation to be wrongly detected as a write.