BLIND_WAIT mode for when the Analog Comparator is busy.
This will be needed if we want to do an ADC read because the Analog
Comparator uses the same multiplexer as the ADC, so either the AC or the
ADC can be used at a time.  It lets the commutation loop skip using the
AC for detecting the zero-crossing and wait blindly for the amount of
time measured in the previous commutation hoping the timing has not
changed drastically.