Let the host read VBAT, add dummy 0x1234 to comm counter values.
This lets the Flight Controller read the last voltage value measured by
the ADC using the serial bitbanging protocol.  The dummy 0x1234 value
that is used by the Flight Controller to check if ESC is alive is now
added to the commutation counter so that if the motor is stopped, it can
read it and check if the ESC is alive and protocol version is what it
expects it to be.  If the motor is running and the FC wants the
commutations count instead, it'll need to subtract 0x1234 from the value
received.