Defer the comparator multiplexer switch until needed.
This allows us to use the ADC whenever we're not using the Analog
Comparator.  Without it we could only use the ADC on the ESC boards that
connect all three phases through the multiplexer (not AIN1) and so only
ADMUX needed to be modified when the phases were switched.  Since ADMUX
is buffered it didn't disturb the ongoing conversion.